Cells store, new buttons
To make the tree appear, you now hit "iterate" and then the "go" button under the LSystem selectore. Boxes have also been added so users can input their own values.
Cellular automata the cells now store, but like who the hell knows if correctly. I am working on implementing a new function that will draw the cells on the screen, so I can clear it and then see if it will draw the 2d array correctly back on the screen. :)