Working on generalization
Added togglebuttons the correspond to a hashmap that correspond to the cellular automata's rules and will EVENTUALLY be able to be fed into a seperate function in cells that will parse the rule sets. Doesn't work yet, and I expect to find problems with conways game of life when I start generalizing it.