Generalized Cellular Automata Try 1
So, this is my attempt at generalizing cellular automata. It does SOMETHING. I am not sure how well it is generalized though and will have to more carefully test it. I also managed to forget that a cell having ZERO neighbors is a potential thing. It does some cool shit though.