UHHHH rainbow?
So there is a list of colors stored in "rainbow" or a hashmap that the program is supposed to iterate over. This is weird though, since they are definitely iterated over, and the automata goes through a "rainbow" but it isn't the dead cells slowly dying going through the rainbow before disapearing, but rather it is alive cells going through it before dying and reproducing. Odd!