Ints now work
Because they are doubles now primarily. But it is cutting things in half appropraitely it looks like, getting F/T/T/T for my example image when it should be getting T/T/T/F. The for loops need to be looked at intensely to understand what it is actually doing.